refactor: API_LIST location/behavior
