+++ src/config/smartapi_overrides.js
@@ -1,0 +1,7 @@
+exports.overrides = {
+  config: {
+    override_API_LIST: true,
+    only_overrides: false,
+  },
+  apis: {}
+}
