--- src/config/smartapi_overrides.json
@@ -1,6 +1,0 @@
-{
-  "conf": {
-    "only_overrides": false
-  },
-  "apis": {}
-}
