+++ src/routes/v1/query_v1.js
@@ -1,5 +1,5 @@
 const path = require("path");
-const config = require("./config");
+const config = require("../../config/apis.js");
 const TRAPIGraphHandler = require("@biothings-explorer/query_graph_handler");
 const swaggerValidation = require("../../middlewares/validate");
 const smartAPIPath = path.resolve(__dirname, process.env.STATIC_PATH ? `${process.env.STATIC_PATH}/data/smartapi_specs.json` : '../../../data/smartapi_specs.json');
